+++ src/assets/main.css
@@ -16,9 +16,3 @@ a,
   color: hsla(160, 100%, 37%, 1);
   transition: 0.4s;
 }
-
-@media (hover: hover) {
-  a:hover {
-    background-color: hsla(160, 100%, 37%, 0.2);
-  }
-}
