+++ src/components/MeetingLog.vue
@@ -5,7 +5,7 @@ import { meetings } from '../assets/history'
 
 <template>
   <h1>Video Game Club Meeting History</h1>
-  <div v-for="meet in meetings">
+  <div v-for="meet in meetings.slice().reverse()">
     <MeetingEntry v-bind="meet"></MeetingEntry>
   </div>
 </template>
